requirements.txt
asgiref==3.4.1
Django==3.2.5
django-background-tasks==1.2.5
django-compat==1.0.15
django-user-agents==0.4.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
ua-parser==0.10.0
user-agents==2.2.0
